Inline test kernel
